Edited this to get working
I can't find anything in the manual about the variable 'SPCM_MAX_AIRANGE' it seems to be a analogue input max limit, but it causes this init code to fail. Commenting this out to get working. Not really sure what difference it would make as it's just to ensure this while loop doesn't repeat beyond max no of analogue inputs. Not tested on other DAQ card.